:art: Moves AppBar into Navigation folder
Condenses file structure by moving AppBar into the Navigation Folder
