+++ client/src/components/Navigation/AppBar/AppBar.js
@@ -15,10 +15,10 @@ import {
 import Avatar from "@material-ui/core/Avatar";
 
 //Logo
-import Logo from "../../img/training-bot.png";
+import Logo from "../../../img/training-bot.png";
 
 //AUTH
-import { login, logout } from "../../Auth/Auth";
+import { login, logout } from "../../../Auth/Auth";
 
 class AppBar extends Component {
   render() {
@@ -38,23 +38,23 @@ class AppBar extends Component {
               <p onClick={() => login()}>Login</p>
             </>
           ) : (
-            <>
-              <Link to="/home">
-                <NavigationLogo alt="A cute, personable robot" src={Logo} />
-              </Link>
-              <RightItemsContainer>
-                <Link to="/home/profile">
-                  {" "}
-                  <Avatar
-                    data-tour="6"
-                    src={JSON.parse(localStorage.getItem("Profile")).picture}
-                    alt={JSON.parse(localStorage.getItem("Profile")).name}
-                  />
+              <>
+                <Link to="/home">
+                  <NavigationLogo alt="A cute, personable robot" src={Logo} />
                 </Link>
-                <LogoutStyling onClick={() => logout()}>Logout</LogoutStyling>
-              </RightItemsContainer>
-            </>
-          )}
+                <RightItemsContainer>
+                  <Link to="/home/profile">
+                    {" "}
+                    <Avatar
+                      data-tour="6"
+                      src={JSON.parse(localStorage.getItem("Profile")).picture}
+                      alt={JSON.parse(localStorage.getItem("Profile")).name}
+                    />
+                  </Link>
+                  <LogoutStyling onClick={() => logout()}>Logout</LogoutStyling>
+                </RightItemsContainer>
+              </>
+            )}
         </NavItemsContainer>
       </NavigationContainer>
     );
